PROG-234 added some path planner sutff
