+++ src/main/deploy/pathplanner/3Ball.path
@@ -1,1 +1,1 @@
-{"waypoints":[{"anchorPoint":{"x":8.695961391924445,"y":1.7099314198589468},"prevControl":null,"nextControl":{"x":8.714249428500517,"y":0.8321056642074357},"holonomicAngle":0.0,"isReversal":false,"velOverride":null,"isLocked":true},{"anchorPoint":{"x":7.8638557277131165,"y":0.4114808229577534},"prevControl":{"x":8.641097282196235,"y":0.40233680466971716},"nextControl":{"x":6.501736762941425,"y":0.4275057519550669},"holonomicAngle":0.0,"isReversal":false,"velOverride":0.1,"isLocked":false},{"anchorPoint":{"x":4.663449326900315,"y":2.1671323342607756},"prevControl":{"x":6.245364490730642,"y":1.5361950723862514},"nextControl":null,"holonomicAngle":0.0,"isReversal":false,"velOverride":null,"isLocked":false}],"maxVelocity":2.0,"maxAcceleration":null,"isReversed":false}
+{"waypoints":[{"anchorPoint":{"x":8.330200660398297,"y":1.9568199136442825},"prevControl":null,"nextControl":{"x":8.348488696969687,"y":1.3624587249302538},"holonomicAngle":0.0,"isReversal":false,"velOverride":null,"isLocked":false},{"anchorPoint":{"x":8.00101600202898,"y":0.8046736093516742},"prevControl":{"x":8.58101523943865,"y":1.248202437959057},"nextControl":{"x":7.6901193802357355,"y":0.5669291338627223},"holonomicAngle":0.0,"isReversal":false,"velOverride":null,"isLocked":false},{"anchorPoint":{"x":5.202946405889787,"y":1.8013716027476614},"prevControl":{"x":5.670694403413219,"y":0.3141728413911049},"nextControl":{"x":4.8463296926563615,"y":2.935229870464196},"holonomicAngle":0.0,"isReversal":false,"velOverride":1.0,"isLocked":false},{"anchorPoint":{"x":1.4721869443708646,"y":1.4264668529381619},"prevControl":{"x":2.6791973583916917,"y":2.5146050292145152},"nextControl":{"x":2.6791973583916917,"y":2.5146050292145152},"holonomicAngle":0.0,"isReversal":true,"velOverride":null,"isLocked":false},{"anchorPoint":{"x":4.572009144015263,"y":2.55118110236666},"prevControl":{"x":3.7124714249398245,"y":2.3134366268777096},"nextControl":null,"holonomicAngle":0.0,"isReversal":false,"velOverride":null,"isLocked":false}],"maxVelocity":4.0,"maxAcceleration":3.0,"isReversed":false}
