+++ src/main/java/com/lightningrobotics/voidrobot/commands/auto/paths/OneBall.java
@@ -4,7 +4,7 @@ import com.lightningrobotics.common.auto.Path;
 import com.lightningrobotics.common.command.core.TimedCommand;
 import com.lightningrobotics.voidrobot.commands.auto.commands.AutonDeployIntake;
 import com.lightningrobotics.voidrobot.commands.auto.commands.AutonIntake;
-import com.lightningrobotics.voidrobot.commands.auto.commands.AutonShootCargo;
+import com.lightningrobotics.voidrobot.commands.auto.commands.AutonShootCargoVision;
 import com.lightningrobotics.voidrobot.commands.turret.AimTurret;
 import com.lightningrobotics.voidrobot.subsystems.*;
 
@@ -35,7 +35,7 @@ public class OneBall extends ParallelCommandGroup {
 			path.getCommand(drivetrain),
 
 			// Shoot 2 (Preload & Collected)
-			new AutonShootCargo(shooter, hood, indexer, turret, vision)
+			new AutonShootCargoVision(shooter, hood, indexer, turret, vision)
 
 		));
 	}
