+++ src/main/java/com/lightningrobotics/voidrobot/commands/auto/paths/TwoBall.java
@@ -4,7 +4,7 @@ import com.lightningrobotics.common.auto.Path;
 import com.lightningrobotics.common.command.core.TimedCommand;
 import com.lightningrobotics.voidrobot.commands.auto.commands.AutonDeployIntake;
 import com.lightningrobotics.voidrobot.commands.auto.commands.AutonIntake;
-import com.lightningrobotics.voidrobot.commands.auto.commands.AutonShootCargo;
+import com.lightningrobotics.voidrobot.commands.auto.commands.AutonShootCargoVision;
 import com.lightningrobotics.voidrobot.commands.turret.AimTurret;
 import com.lightningrobotics.voidrobot.subsystems.*;
 
@@ -39,7 +39,7 @@ public class TwoBall extends ParallelCommandGroup {
 			),
 
 			// Shoot 2 (Preload & Collected)
-			new AutonShootCargo(shooter, hood, indexer, turret, vision)
+			new AutonShootCargoVision(shooter, hood, indexer, turret, vision)
 
 		));
 	}
